fixes issue #4485
Replacing the plugin `wsdjeg/vim-lisp` to `vlime/vlime` . The former repository is totally outdated and does not even work while the the later one is actively maintained and contains many more function. In this commit, only fixes are applied to make the key bindings compatible with the `Vlime` API.

In future more Vlime specific functions should(would) be implemented with SpaceVim style keybindings. As of now all Vlime default keybindings work perfectly as expected.